refactor : count 반환 타입 Integer -> Long 변경

+++ src/main/java/com/owori/domain/heart/repository/HeartRepository.java
@@ -12,5 +12,5 @@ public interface HeartRepository {
     Heart save(Heart heart);
     Optional<Heart> findByMemberAndStory(Member member, Story story);
     boolean existsByMemberAndStory(Member member, Story story);
-    Integer countByMember(Member member);
+    Long countByMember(Member member);
 }
